Update of the example
The new dll is tested, it can read energy value properly. Now the example works well, except that if I connect a power sensor and an energy sensor at the same time, the power value won't refresh unless a pulse is detected by the energy sensor. It seems that the program will be stuck at measPower and measEnergy methods until the values are got successfully? I'm not very sure.
This won't influence the usage of the example for most cases.